common, include: Consolidates simple_spin.h and Spinlock.h; removes ceph_spinlock_t; implements Spinlock with std::atomic_flag.
Moves custom atomic implementations to the C++ standard library.
