+++ src/msg/simple/SimpleMessenger.cc
@@ -54,7 +54,6 @@ SimpleMessenger::SimpleMessenger(CephContext *cct, entity_name_t name,
 {
   ANNOTATE_BENIGN_RACE_SIZED(&timeout, sizeof(timeout),
                              "SimpleMessenger read timeout");
-  ceph_spin_init(&global_seq_lock);
   init_local_connection();
 }
 
@@ -67,7 +66,6 @@ SimpleMessenger::~SimpleMessenger()
   assert(!did_bind); // either we didn't bind or we shut down the Accepter
   assert(rank_pipe.empty()); // we don't have any running Pipes.
   assert(!reaper_started); // the reaper thread is stopped
-  ceph_spin_destroy(&global_seq_lock);
 }
 
 void SimpleMessenger::ready()
